Permission controller, seeder, blade

+++ app/Http/Controllers/PermissionController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use App\Models\Permission;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Validation\Rule;
 use Illuminate\View\View;
 
 class PermissionController extends Controller
@@ -14,6 +16,7 @@ class PermissionController extends Controller
     //@route GET /permissions
     public function index():View
     {
+        Gate::authorize('have-permission', ['user-management']);
         $permissions = Permission::paginate(10);
         return view('permissions.index')->with(['permissions' => $permissions]);
     }
@@ -22,6 +25,7 @@ class PermissionController extends Controller
     //@route POST /permissions
     public function store(Request $request):RedirectResponse
     {
+        Gate::authorize('have-permission', ['user-management']);
         $data = $request->validate([
             'name' => 'required|unique:permissions'
         ]);
@@ -35,13 +39,40 @@ class PermissionController extends Controller
     //@route GET /permissions/create
     public function create():View
     {
+        Gate::authorize('have-permission', ['user-management']);
         return view('permissions.create');
+    }
+
+    //@desc Show edit permission form
+    //@route GET /permissions/{permission}/edit
+    public function edit(Permission $permission): View
+    {
+        Gate::authorize('have-permission', ['user-management']);
+        return view('permissions.edit')->with(['permission' => $permission]);
+    }
+
+    //@desc Update  permission details
+    //@route POST /permissions/{permission}/edit
+    public function update(Request $request,Permission $permission): RedirectResponse
+    {
+        Gate::authorize('have-permission', ['user-management']);
+        $data = $request->validate([
+            'name' => [
+                'required',
+                Rule::unique('permissions')->ignore($permission->id),
+            ]
+        ]);
+
+        $permission->update($data);
+
+        return redirect()->route('permissions.index')->with(['success' => 'Permission successfully updated.']);
     }
 
     //@desc Delete permission
     //@route DELETE /permissions
     public function destroy(Permission $permission)
     {
+        Gate::authorize('have-permission', ['user-management']);
         $permission->delete();
         return redirect()->route('permissions.index')->with('success', 'Permission deleted successfully.');
     }
